[수정] 바나나 모델 코드 리팩토링
