+++ SmartAI/Models/ChartInfo.swift
@@ -1,32 +1,23 @@
-////
-////  ChartInfo.swift
-////  SmartAI
-////
-////  Created by Hamlit Jason on 2022/11/22.
-////
 //
-//import Foundation
+//  ChartInfo.swift
+//  SmartAI
 //
-//class ChartItem: Identifiable {
-//    let id = UUID().uuidString
-//    
-//    var bananaClass: String
-//    var probability: String
-//    
-//    init(bananaClass: String, probability: String) {
-//        self.bananaClass = bananaClass
-//        self.probability = probability
-//    }
-//}
+//  Created by Hamlit Jason on 2022/11/22.
 //
-//class ChartInfo: ObservableObject, Identifiable {
-//    let id = UUID().uuidString
-//    
-//    @Published var localData: [ChartItem] = [] {
-//        didSet { print("🧓🏾 \(localData)")}
-//    }
-//    
-//    @Published var serverData: [ChartItem] = [] {
-//        didSet { print("🧕🏻 \(serverData)")}
-//    }
-//}
+
+import Foundation
+
+// MARK: - 바나나 차트
+struct BananaChartInfo: Identifiable {
+    var id = UUID().uuidString
+    
+    /**바나나 이름 */ var name: String
+    /** 확률 */ var probability: String
+}
+
+struct ChartInfo: Identifiable {
+    let type: String
+    let bananas: [BananaChartInfo]
+    
+    var id: String { type }
+}
